Ajustes Migração CDR

+++ WimDesktop/Views/MigrationDatabaseView.cs
@@ -1,16 +1,17 @@
-﻿using WimDesktop._Repositories;
-using WimDesktop.Interface.IRepository;
-using WimDesktop.Models;
+﻿using FellowOakDicom;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Windows.Forms;
 using System.Linq;
-using System.Threading.Tasks;
-using WimDesktop.Properties;
 using System.Text;
 using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+using WimDesktop._Repositories;
+using WimDesktop.Interface.IRepository;
 using WimDesktop.Interface.IView;
+using WimDesktop.Models;
+using WimDesktop.Properties;
 
 namespace WimDesktop.Views
 {
@@ -19,7 +20,6 @@ namespace WimDesktop.Views
         private readonly string wimMigrationPath = @"C:\WimDesktopDB\migration\tools\";
         private readonly string dataPath = "";
         private string migrationErrors = "";
-
 
         public string software { get; set; }
         public List<PatientModel> patients { get; set; } = new List<PatientModel>();
@@ -171,76 +171,90 @@ namespace WimDesktop.Views
 
         private async Task importData_CDR(IProgress<int> progress, int processedPatients)
         {
-            int totalPatients = patients.Count;
-
-            foreach (PatientModel patient in patients)
+            try
             {
-                int patientOldId = patient.id;
+                int totalPatients = patients.Count;
 
-                List<ExamModel> patientExams = exams.Where(e => e.patientId == patient.id).ToList();
-
-                patientToImport = patient;
-                eventImportPatient?.Invoke(this, EventArgs.Empty);
-
-                if (patientExams.Any())
+                foreach (PatientModel patient in patients)
                 {
-                    foreach (ExamModel patientExam in patientExams)
+                    int patientOldId = patient.id;
+
+                    List<ExamModel> patientExams = exams.Where(e => e.patientId == patient.id).ToList();
+
+                    patientToImport = patient;
+                    eventImportPatient?.Invoke(this, EventArgs.Empty);
+
+                    if (patientExams.Any())
                     {
-                        int examOldId = patientExam.id;
-                        patientExam.patientId = patient.id;
-
-                        List<ExamImageCDR> currentExamImages = examImagesCDR.Where(ei => ei.examId == patientExam.id).ToList();
-
-                        if (!currentExamImages.Any())
+                        foreach (ExamModel patientExam in patientExams)
                         {
-                            continue;
-                        }
+                            int examOldId = patientExam.id;
 
-                        string filePath = currentExamImages.First().file;
-                        int index = filePath.LastIndexOf('\\');
+                            patientExam.patientId = patient.id;
 
-                        string examContentPath = Path.Combine(dataPath, filePath.Substring(0, index), "SR000001");
-
-                        if (File.Exists(examContentPath))
-                        {
-                            currentExamImages = getUsedImages(examContentPath, currentExamImages);
+                            List<ExamImageCDR> currentExamImages = examImagesCDR.Where(ei => ei.examId == patientExam.id).ToList();
 
                             if (!currentExamImages.Any())
                             {
                                 continue;
                             }
-                        }
-                        else
-                        {
-                            examContentPath = Path.Combine(dataPath, filePath);
-                        }
 
+                            string filePath = currentExamImages.First().file;
+                            string dirPath = Path.Combine(dataPath, Path.GetDirectoryName(filePath));
 
-                        if (currentExamImages.Any())
-                        {
-                            patientExam.templateId = getTemplateId(examContentPath, currentExamImages.Count);
-
-                            if (patientExam.templateId == 0)
+                            if (Directory.Exists(dirPath) && !Directory.GetFiles(dirPath, "*.jpeg").Any())
                             {
                                 continue;
                             }
 
-                            examToImport = patientExam;
-                            eventImportExam?.Invoke(this, EventArgs.Empty);
+                            string examContentPath = Path.Combine(dirPath, "SR000001");
 
-                            getTemplateFrameId(currentExamImages, patientExam.templateId);
+                            if (File.Exists(examContentPath))
+                            {
+                                currentExamImages = getUsedImages(examContentPath, currentExamImages);
 
-                            await importImages_CDR(currentExamImages, patient.id, patientExam.id);
+                                if (!currentExamImages.Any())
+                                {
+                                    continue;
+                                }
+                            }
+                            else
+                            {
+                                examContentPath = Path.Combine(dataPath, filePath);
+                            }
 
-                            List<ExamImageModel> patientExamImages = convertExamImageModels(currentExamImages, patientExam.id);
+                            if (currentExamImages.Any())
+                            {
+                                patientExam.templateId = getTemplateId(examContentPath, currentExamImages.Count);
 
-                            examImagesToImport = patientExamImages;
-                            eventImportExamImages?.Invoke(this, EventArgs.Empty);
+                                if (patientExam.templateId == 0)
+                                {
+                                    continue;
+                                }
+
+                                examToImport = patientExam;
+                                eventImportExam?.Invoke(this, EventArgs.Empty);
+
+
+                                getTemplateFrameId(currentExamImages, patientExam.templateId);
+
+                                await importImages_CDR(currentExamImages, patient.id, patientExam.id);
+
+                                List<ExamImageModel> patientExamImages = convertExamImageModels(currentExamImages, patientExam.id);
+
+                                examImagesToImport = patientExamImages;
+
+                                eventImportExamImages?.Invoke(this, EventArgs.Empty);
+                            }
                         }
                     }
+                    processedPatients++;
+                    progress.Report((processedPatients * 100) / totalPatients);
                 }
-                processedPatients++;
-                progress.Report((processedPatients * 100) / totalPatients);
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message + ex.StackTrace);
             }
         }
 
@@ -260,17 +274,16 @@ namespace WimDesktop.Views
         {
             string content = File.ReadAllText(path, Encoding.Default);
 
-            var lines = content.Split('à').Where(p => p.Contains("SCHICK TECHNOLOGIES"));
+            var matches = Regex.Matches(content, @"\b\d+(\.\d+)+\b");
 
-            List<string> imagesIdentifier = new List<string>();
+            HashSet<string> imagesIdentifier = new HashSet<string>(); // evita duplicados
 
-            foreach (var line in lines)
+            foreach (Match match in matches)
             {
-                Match match = Regex.Matches(line, @"\d+(\.\d+)*").Cast<Match>().FirstOrDefault(m => m.Length > 2);
-
-                if (match != null)
+                string uid = match.Value.Trim();
+                if (!string.IsNullOrWhiteSpace(uid))
                 {
-                    imagesIdentifier.Add(match.Value);
+                    imagesIdentifier.Add(uid);
                 }
             }
 
@@ -292,44 +305,35 @@ namespace WimDesktop.Views
             {
                 string content = File.ReadAllText(path, Encoding.Default);
 
-                int init = content.IndexOf("Sirona Dental");
-                int final = content.IndexOf("CDR");
+                string cleanContent = Regex.Replace(content, @"[^\x20-\x7E]", "");
 
-                string text = content.Substring(init, final - init);
+                string[] keywords = new string[] { "BITEWINGS", "CENTRAL INFERIOR", "CENTRAL SUPERIOR", "CHECKUP", "ENDO HORIZONTAL (15)", "FULL MOUTH SERIES (18)", "FULL MOUTH SERIES (21)", "MANDIBULAR DIREITO", "MANDIBULAR ESQUERDO", "MAXILAR DIREITO", "MAXILAR DIREITO", "VERTICAL ENDO" };
 
-                string resultClean = Regex.Replace(text, @"[^\x20-\x7E]", "");
-
-                string[] parts = resultClean.Split(new string[] { "LO" }, StringSplitOptions.None);
-
-                ITemplateRepository templateRepository = new TemplateRepository();
-
-                List<TemplateModel> templates = templateRepository.getAllTemplates();
-
-                TemplateModel matchedTemplate = findMatch(parts, templates);
-
-                if (matchedTemplate != null)
+                foreach (string keyword in keywords)
                 {
-                    return matchedTemplate.id;
+                    if (cleanContent.Contains(keyword))
+                    {
+                        ITemplateRepository templateRepository = new TemplateRepository();
+                        List<TemplateModel> templates = templateRepository.getAllTemplates();
+
+                        TemplateModel matchedTemplate = findMatch(new string[] { keyword }, templates);
+                        if (matchedTemplate != null)
+                        {
+                            return matchedTemplate.id;
+                        }
+                    }
                 }
+
+                if (imagesCount < 6)
+                    return 13;
+                else if (imagesCount < 16)
+                    return 14;
                 else
-                {
-                    if (imagesCount < 6)
-                    {
-                        return 13;
-                    }
-                    else if (imagesCount < 16)
-                    {
-                        return 14;
-                    }
-                    else
-                    {
-                        return 15;
-                    }
-                }
+                    return 15;
             }
             catch (Exception ex)
             {
-                migrationErrors += "" + ex.Message + "\n\n";
+                migrationErrors += ex.Message + "\n\n";
                 return 0;
             }
         }
@@ -417,13 +421,18 @@ namespace WimDesktop.Views
 
             Directory.CreateDirectory(Path.Combine(folderDest, "recycle"));
 
-            foreach (var examImage in examImages)
+            foreach (ExamImageCDR examImage in examImages)
             {
                 string originFile = Path.Combine(dataPath, examImage.file + ".jpeg");
 
-                string dest = Path.Combine(folderDest, $"{examImage.order}_original.png");
+                if (File.Exists(originFile))
+                {
+                    string dest = Path.Combine(folderDest, $"{examImage.order}_original.png");
 
-                await Task.Run(() => File.Copy(originFile, dest, overwrite: true));
+                    await Task.Run(() => File.Copy(originFile, dest, overwrite: true));
+
+                    examImage.file = dest;
+                }
             }
         }
     }
